Move GAPI into shared features
