--- src/gapi/index.ts
@@ -1,2 +1,0 @@
-export * from './auth';
-export { default as GapiAuthContext } from './context';
