+++ src/profile/components/connect/Selector.tsx
@@ -1,7 +1,11 @@
 import styled from 'styled-components';
 
-import { handleGoogleLogIn } from '../../../gapi';
-import { NarrowColouredButton, NarrowLightLink, routes } from '../../../shared';
+import {
+  handleGoogleLogIn,
+  NarrowColouredButton,
+  NarrowLightLink,
+  routes,
+} from '../../../shared';
 
 interface Props {
   onSelectCsv: () => void;
