+++ src/shared/index.ts
@@ -7,5 +7,7 @@ export { default as GlobalContextProvider } from './components/GlobalContextProv
 export { default as Image } from './components/Image';
 export { default as Input } from './components/Input';
 export { default as PageWrapper } from './components/PageWrapper';
+export * from './gapi/auth';
+export { default as GapiAuthContext } from './gapi/context';
 export { logInfo, logWarning, logError } from './reporting';
 export * as routes from './routes';
